Fix test configuration of 'generate' command
